# Optimize CUDA Sum op kernel and refactor CUDA elementwise variadic input op kernels