**Description**
For the special case where all variadic inputs of a kernel are the same shape (i.e. no broadcasting is required) and there are few enough of them, we can perform the entire computation in a single kernel. The general implementation (which was previously used for this special case) handles broadcasting by repeatedly invoking a binary kernel on successive inputs.

**Motivation and Context**
Performance. The original motivation was to optimize Sum, but after some refactoring Min and Max should get similar benefits.

**Performance tests**
baseline (0349479b19d7735956be3cfb0bc1c8014a214dbb), updated (04f27de6c95d525a0ba2cab672ed44b495526a6d)

CUDA kernel benchmark:
The CUDA Sum op kernel was run 100 times with the specified input shape and number of inputs.
num_calls, average_duration, total_time and total_percentage refer to CUDA kernel calls.
The particular case motivating the optimization has four inputs, shown here. The full data including other input counts can be found in a separate PR comment.

baseline:
input_shape|input_count|num_calls|average_duration|total_time|total_percentage
-|-|-|-|-|-
[1024]|4|400|1.570000us|628.221000us|36.347473
[1024,1024]|4|400|13.243000us|5.297296ms|2.439728
[16,1024,1024]|4|400|0.238441ms|0.095377s|2.263378

updated:
input_shape|input_count|num_calls|average_duration|total_time|total_percentage
-|-|-|-|-|-
[1024]|4|100|2.167000us|216.700000us|19.038917
[1024,1024]|4|100|30.299000us|3.029963ms|1.520693
[16,1024,1024]|4|100|0.416080ms|0.041608s|1.083528

BERT large throughput:
```
./onnxruntime_training_bert --model_name /bert_ort/bert_models/nv/bert-large/bert-large-uncased_L_24_H_1024_A_16_V_30528_S_512_Dp_0.1_optimized_layer_norm --train_data_dir /bert_data/128/books_wiki_en_corpus/train --test_data_dir /bert_data/128/books_wiki_en_corpus/test --train_batch_size 64 --mode train --num_train_steps 400 --display_loss_steps 1 --warmup_ratio=0.2843 --warmup_mode=Poly --optimizer lamb --gradient_accumulation_steps 1 --max_predictions_per_seq=20 --use_nccl --use_mixed_precision --allreduce_in_fp16
```

baseline:
Stabilized Throughput: 180.306 Examples / Second

updated:
Stabilized Throughput: 182.129 Examples / Second

1.01% improvement