@edgchen1 2020-07-06
CUDA kernel benchmark:
The CUDA Sum op kernel was run 100 times with the specified input shape and number of inputs.
num_calls, average_duration, total_time and total_percentage refer to CUDA kernel calls.

baseline:
input_shape|input_count|num_calls|average_duration|total_time|total_percentage
-|-|-|-|-|-
[1024]|2|100|1.558000us|155.839000us|22.163517
[1024]|3|300|1.552000us|465.885000us|33.835373
[1024]|4|400|1.570000us|628.221000us|36.347473
[1024]|5|500|1.544000us|772.347000us|37.608996
[1024]|6|600|1.546000us|0.927672ms|38.791988
[1024]|7|700|1.560000us|1.092666ms|39.853914
[1024]|8|800|1.565000us|1.252438ms|40.639953
[1024]|9|900|1.550000us|1.395730ms|40.991770
[1024]|10|1000|1.555000us|1.555218ms|41.493304
[1024]|11|1100|1.553000us|1.708400ms|41.782175
[1024]|12|1200|1.575000us|1.890648ms|42.465583
[1024]|13|1300|1.547000us|2.011733ms|42.349528
[1024]|14|1400|1.553000us|2.175374ms|42.653171
[1024]|15|1500|1.559000us|2.339481ms|42.939202
[1024]|16|1600|1.568000us|2.510278ms|43.254561
[1024,1024]|2|100|16.078000us|1.607825ms|1.243748
[1024,1024]|3|300|13.252000us|3.975843ms|2.368285
[1024,1024]|4|400|13.243000us|5.297296ms|2.439728
[1024,1024]|5|500|13.315000us|6.657829ms|2.630315
[1024,1024]|6|600|13.298000us|7.978875ms|2.653499
[1024,1024]|7|700|13.344000us|9.341331ms|2.769100
[1024,1024]|8|800|13.355000us|10.684610ms|2.328100
[1024,1024]|9|900|13.358000us|12.022903ms|2.484503
[1024,1024]|10|1000|13.361000us|13.361063ms|2.652611
[1024,1024]|11|1100|13.351000us|14.686674ms|2.105765
[1024,1024]|12|1200|13.386000us|16.063511ms|2.590635
[1024,1024]|13|1300|13.364000us|17.373331ms|2.094010
[1024,1024]|14|1400|13.370000us|18.718939ms|2.613680
[1024,1024]|15|1500|13.357000us|20.036491ms|2.569900
[1024,1024]|16|1600|13.365000us|21.384485ms|2.617951
[16,1024,1024]|2|100|0.242198ms|0.024220s|0.793774
[16,1024,1024]|3|300|0.238494ms|0.071548s|2.164877
[16,1024,1024]|4|400|0.238441ms|0.095377s|2.263378
[16,1024,1024]|5|500|0.238443ms|0.119222s|2.289291
[16,1024,1024]|6|600|0.238446ms|0.143068s|2.418229
[16,1024,1024]|7|700|0.238425ms|0.166898s|2.410617
[16,1024,1024]|8|800|0.238380ms|0.190704s|2.454700
[16,1024,1024]|9|900|0.238414ms|0.214573s|2.289505
[16,1024,1024]|10|1000|0.238412ms|0.238412s|2.496864
[16,1024,1024]|11|1100|0.238415ms|0.262257s|2.544768
[16,1024,1024]|12|1200|0.238421ms|0.286106s|2.389313
[16,1024,1024]|13|1300|0.238402ms|0.309923s|2.560218
[16,1024,1024]|14|1400|0.238370ms|0.333719s|2.511502
[16,1024,1024]|15|1500|0.238384ms|0.357576s|2.492723
[16,1024,1024]|16|1600|0.238384ms|0.381415s|2.571946

updated:
input_shape|input_count|num_calls|average_duration|total_time|total_percentage
-|-|-|-|-|-
[1024]|2|100|1.552000us|155.261000us|22.177418
[1024]|3|100|2.948000us|294.844000us|28.767457
[1024]|4|100|2.976000us|297.691000us|24.541512
[1024]|5|100|2.995000us|299.549000us|21.478482
[1024]|6|100|3.031000us|0.303192ms|19.136910
[1024]|7|100|3.055000us|0.305501ms|17.238770
[1024]|8|100|3.101000us|0.310142ms|15.855006
[1024]|9|100|3.205000us|0.320569ms|14.927571
[1024]|10|100|3.298000us|0.329853ms|14.082705
[1024]|11|100|3.427000us|0.342750ms|13.504249
[1024]|12|100|3.600000us|0.360028ms|13.142135
[1024]|13|100|3.786000us|0.378685ms|12.893090
[1024]|14|100|4.011000us|0.401117ms|12.746478
[1024]|15|100|4.207000us|0.420760ms|12.570999
[1024]|16|100|4.421000us|0.442174ms|12.428522
[1024,1024]|2|100|15.995000us|1.599540ms|1.179677
[1024,1024]|3|100|24.969000us|2.496998ms|1.423353
[1024,1024]|4|100|31.676000us|3.167655ms|1.452848
[1024,1024]|5|100|37.708000us|3.770859ms|1.480264
[1024,1024]|6|100|43.392000us|4.339231ms|1.486333
[1024,1024]|7|100|49.233000us|4.923377ms|1.432832
[1024,1024]|8|100|55.007000us|5.500778ms|1.386525
[1024,1024]|9|100|60.726000us|6.072652ms|1.333576
[1024,1024]|10|100|66.004000us|6.600476ms|1.293678
[1024,1024]|11|100|71.610000us|7.161051ms|1.222433
[1024,1024]|12|100|86.225000us|8.622585ms|1.173041
[1024,1024]|13|100|83.865000us|8.386519ms|1.182757
[1024,1024]|14|100|89.106000us|8.910606ms|1.180123
[1024,1024]|15|100|96.259000us|9.625929ms|1.213998
[1024,1024]|16|100|100.489000us|10.048938ms|1.178890
[16,1024,1024]|2|100|0.242201ms|0.024220s|0.906992
[16,1024,1024]|3|100|0.330688ms|0.033069s|0.948985
[16,1024,1024]|4|100|0.416721ms|0.041672s|0.957807
[16,1024,1024]|5|100|0.500636ms|0.050064s|0.953149
[16,1024,1024]|6|100|0.584466ms|0.058447s|0.946830
[16,1024,1024]|7|100|0.667572ms|0.066757s|0.963874
[16,1024,1024]|8|100|0.753688ms|0.075369s|0.948259
[16,1024,1024]|9|100|0.832439ms|0.083244s|0.900331
[16,1024,1024]|10|100|0.918186ms|0.091819s|0.899297
[16,1024,1024]|11|100|0.997243ms|0.099724s|0.916204
[16,1024,1024]|12|100|1.079014ms|0.107901s|0.907696
[16,1024,1024]|13|100|1.203035ms|0.120304s|0.916146
[16,1024,1024]|14|100|1.244234ms|0.124423s|0.857107
[16,1024,1024]|15|100|1.389488ms|0.138949s|1.068000
[16,1024,1024]|16|100|1.412288ms|0.141229s|0.886164